requirements.txt
fastapi
uvicorn
groq
python-dotenv
pymupdf
sqlalchemy
python-multipart
